[search] Call detrivialise during normalisation process only at root
detrivialise can remove the current node if the current node is redundant,
it will attach the child nodes to the parent node directly. Thus, it is
not good to run detrivialise in the middle of the tree as it can cause
dangeling or NULL pointers.

Share the nnodes_last_collect variable across the normalise call
hierarchy and restart the root normalisation process when a detrivialise
is needed.

Signed-off-by: Mildred Ki'Lya <mildred-pub.git@mildred.fr>